Add a function for checking synchronized output support
This change adds a function that queries the terminal to check whether
it supports the synchronized output feature (see e065a56 and
https://gist.github.com/christianparpart/d8a62cc1ab659194337d73e399004036).

This query works the same as the query for Kitty's keyboard enhancement
protocol (see 3fe13e1). We send a query sequence for the synchronized
output mode-setting and the query sequence for the primary device
attributes and then poll for a response to  either. The terminal should
return a DECRPM report sequence in the shape 'CSI ? 2026 ; <n> $ y'
where `<n>` is 1 or 2 to signal support and 0 or 4 for not supported.
The terminal may not return a response for the synchronized output query
if it doesn't support DECRPM mode-setting reports, so if the primary
device attributes (which are widely supported) arrive, we know that the
terminal does not support synchronized output.